% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{tel_dor}
\alias{akko}
\alias{arad}
\alias{ein_hofez}
\alias{eshtemoa}
\alias{ref_data}
\alias{model}
\title{LIA points of Silver Hoards from Israel}
\format{
**For artefact sampels:** Data frame with 3 variables
\describe{
  \item{pb64}{Lead Isotope ratio of 206Pb/204Pb}
  \item{pb74}{Lead Isotope ratio of 207Pb/204Pb}
  \item{pb84}{Lead Isotope ratio of 208Pb/204Pb}
}

**For reference data:** Object of class `ref.data` with 4 variables.
\describe{
  \item{group}{Regions of Ore samples}
  \item{pb64}{Lead Isotope ratio of 206Pb/204Pb}
  \item{pb74}{Lead Isotope ratio of 207Pb/204Pb}
  \item{pb84}{Lead Isotope ratio of 208Pb/204Pb}
}

**For Xgboost model:** List with xgb.Booster objects

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 3 columns.

An object of class \code{ref.data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7267 rows and 4 columns.

An object of class \code{list} of length 157.
}
\source{
Eshel, T., Erel, Y., Yahalom-Mack, N., Tirosh, O., & Gilboa, A.
  (2019). Lead isotopes in silver reveal earliest Phoenician quest for metals
  in the west Mediterranean. Proceedings of the National Academy of Sciences,
  116(13), 6007–6012.
  \href{https://doi.org/10.1073/pnas.1817951116}{10.1073/pnas.1817951116}
}
\usage{
tel_dor

akko

arad

ein_hofez

eshtemoa

ref_data

model
}
\description{
Lead Isotope ratios of a Phoenician Silver Hoard from Tel Dor, Akko, Arad, Ein Hofez and  Eshtemona.
Reference Data set of Isotope rations and a pretrained model using default parameters from [train_data()]
}
\examples{
# Create object with class liaendmembers
data("tel_dor")
end_members <- endmembers(
        tel_dor,
        colnames(tel_dor),
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
# Print summary of the liaendmembers object
summary(end_members)
# Get Uclidian distance
isoprov_dist(end_members, ref_data, dist_type = "ed")
# Get Mass Fractionnation Corrected distance
isoprov_dist(end_members, ref_data, dist_type = "mfd")
# Get XGBOOST predicted resutls
isoprov_predict(end_members, model)
}
\keyword{datasets}
