% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dor.R
\docType{data}
\name{dor}
\alias{dor}
\title{LIA points from Tel Dor}
\format{
Data frame with 32 observations and 3 variables
\describe{
  \item{pb64}{Lead Isotope ratio of 206Pb/204Pb}
  \item{pb74}{Lead Isotope ratio of 207Pb/204Pb}
  \item{pb84}{Lead Isotope ratio of 208Pb/204Pb}
}
}
\source{
Eshel, T., Erel, Y., Yahalom-Mack, N., Tirosh, O., & Gilboa, A. (2019). Lead isotopes in silver reveal earliest Phoenician quest for metals in the west Mediterranean. Proceedings of the National Academy of Sciences, 116(13), 6007–6012. \href{https://doi.org/10.1073/pnas.1817951116}{10.1073/pnas.1817951116}
}
\usage{
dor
}
\description{
Lead Isotope ratios of a Silver Hoard from Tel Dor
}
\examples{
# Create object with class liaendmembers
data("dor")
end_members <- endmembers(
        dor,
        colnames(dor),
        tolerance = 0.01
)
# Print summary of the liaendmembers object
summary.liaendmembers(end_members)
}
\keyword{datasets}
