% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_frac_euc_dist.R
\name{euc_dist}
\alias{euc_dist}
\alias{mf_dist}
\alias{endmember_dist}
\title{Euclidean Distance for Pb Isotope rations to ore Sources}
\usage{
euc_dist(x, ox, ox_col, group, .n = 10, as.dist = FALSE, ...)

mf_dist(x, ox, ox_col, group, .n = 10, as.dist = FALSE, s = 0.001, ...)

endmember_dist(
  df,
  ox,
  ox_col,
  group,
  dist_type = "ed",
  .n = 10,
  as.dist = FALSE,
  s = 0.001,
  ...
)
}
\arguments{
\item{x}{matrix of Pb Isotopes with colums in the order of
206Pb/204Pb, 207Pb/204Pb,208Pb/204Pb}

\item{ox}{dataframe of renfrence data contained the Pb Isotope rations}

\item{ox_col}{vector contained names of the Pb Isotope rations.
Names much include digits 6, 7, 8 for isotope ratio identification.}

\item{group}{groups belonging to reach isotope ration vector}

\item{.n}{Length of result output (Default = 10)}

\item{as.dist}{Option to show distances or only a vector of closest groups.}

\item{...}{}

\item{s}{mass-fractionation factor (Defualt = 0.001)}

\item{dist_type}{Distance type to use, simple euclidiant or
mass-fractionation corrected}
}
\value{
data frame or charecter vector

list of dataframe or charecter vector
}
\description{
This function
}
