% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoprov_dist.R
\name{isoprov_dist}
\alias{isoprov_dist}
\title{Euclidean Distance for Pb Isotope rations to ore Sources}
\usage{
isoprov_dist(
  x,
  ref,
  dist_type = stop("Distance type should be Defined 'ed' or 'mfd'"),
  .n = 1,
  s = 0.001,
  ...
)
}
\arguments{
\item{x}{Matrix, data frame of `liaendmembers objectof Pb Isotopes with colums in the order of
206Pb/204Pb, 207Pb/204Pb,208Pb/204Pb or a `liaendmember` object.}

\item{ref}{`ref.data` object used for Reference alaysis.}

\item{dist_type}{Distance type to use, simple euclidiant ('ed') or
mass-fractionation corrected ('mfd')}

\item{.n}{Length of result output (Default = 1)}

\item{s}{Mass-fractionation factor (Defualt = 0.001)}

\item{...}{Additional params}
}
\value{
List of dataframe or charecter vector
}
\description{
Calcutalte the culidian distanc of each isotope sample to a reference dataset,
and gives the clossest regions to the groups.
Mass-fractionation follows the procedure outlined in Albarede et.al (2024)
}
\references{
Albarede, F., Davis, G., Blichert-Toft, J., Gentelli, L., Gitler, H., Pinto, M., & Telouk, P. (2024). A new algorithm for using Pb isotopes to determine the provenance of bullion in ancient Greek coinage. Journal of Archaeological Science, 163, 105919. https://doi.org/10.1016/j.jas.2023.105919
}
