% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_data.R
\name{ref_data}
\alias{ref_data}
\title{Cereate Reference data object for LIA endmember distance and probablity estimate functions.}
\usage{
ref_data(x, cols, group)
}
\arguments{
\item{x}{Reference data frame or tibble.}

\item{cols}{Nector of column names which represent 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb}

\item{group}{Name of the column containing siotope groups as character string.}
}
\value{
ref.data object with Isotope groupings and Isotope colums of
206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb
with shorter names.
}
\description{
Cereate Reference data object for LIA endmember distance and probablity estimate functions.
}
