% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoprov_predict.R
\name{train_data}
\alias{train_data}
\title{Train XGBOOST Model}
\usage{
train_data(
  ref,
  .minSize = 3,
  .minPts_fac = 0.1,
  .eps = 0.18,
  .eta = 0.1,
  .max_depth = 6,
  .nrounds = 100,
  nthread = 4L,
  ...
)
}
\arguments{
\item{ref}{`ref.data` object created by ref_data.}

\item{.minSize}{Minisume number of samples in group to be used for sampeling (Default = 3).}

\item{.minPts_fac}{scaling factor minimum needed points from each group, ranging from 0:1. (Default = 0.1)}

\item{.eps}{size (radius) of the epsilon neighborhood. (Default = 0.18)}

\item{.eta}{Step size shrinkage used in update to prevent overfitting. After each boosting step, we can directly get the weights of new features, and eta shrinks the feature weights to make the boosting process more conservative. (Defualt = 0.1)}

\item{.max_depth}{Maximum depth of a tree. Increasing this value will make the model more complex and more likely to overfit. 0 indicates no limit on depth. Beware that XGBoost aggressively consumes memory when training a deep tree. "exact" tree method requires non-zero value. (Default = 6)}

\item{.nrounds}{Max number of boosting iterations. (Default = 100)}

\item{nthread}{Number of threadts for parallel processing. When choosing it, please keep thread contention and hyperthreading in mind. (Default = 4)}

\item{...}{
  Arguments passed on to \code{\link[dbscan:dbscan]{dbscan::dbscan}}, \code{\link[xgboost:xgb.train]{xgboost::xgb.train}}
  \describe{
    \item{\code{weights}}{numeric; weights for the data points. Only needed to perform
weighted clustering.}
    \item{\code{borderPoints}}{logical; should border points be assigned to clusters.
The default is \code{TRUE} for regular DBSCAN. If \code{FALSE} then border
points are considered noise (see DBSCAN* in Campello et al, 2013).}
    \item{\code{early_stopping_rounds}}{Number of boosting rounds after which training will be stopped
if there is no improvement in performance (as measured by the evaluatiation metric that is
supplied or selected by default for the objective) on the evaluation data passed under
\code{evals}.

Must pass \code{evals} in order to use this functionality. Setting this parameter adds the
\code{\link[xgboost:xgb.cb.early.stop]{xgb.cb.early.stop()}} callback.

If \code{NULL}, early stopping will not be used.}
    \item{\code{maximize}}{If \code{feval} and \code{early_stopping_rounds} are set, then this parameter must be set as well.
When it is \code{TRUE}, it means the larger the evaluation score the better.
This parameter is passed to the \code{\link[xgboost:xgb.cb.early.stop]{xgb.cb.early.stop()}} callback.}
  }}
}
\value{
List of xgboot.model objects.
}
\description{
Trains XGBOOST model for predictive isotope analysis using DCSCAN clustring
SMOTE data imputation for training data set, and XGBOOST, following the method
of Shnyr et.al (2026)
}
\references{
Shnyr, E., Kuflik, T., Desai, K., & Eshel, T. (2026). Determining the origins of Phoenician silver: Exploring the potential of machine learning for lead isotope analysis. Journal of Archaeological Science, 188, 106–499. https://doi.org/10.1016/j.jas.2026.106499
}
