% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_fraction_uclidian_distance.R
\name{euc_dist}
\alias{euc_dist}
\alias{mass_frac_dist}
\title{Euclidan and Mass Fractination Distances for Pb isotopes}
\usage{
euc_dist(df, ox, ox_col, group, .n = 10, as.dist = FALSE, ...)

mass_frac_dist(df, ox, ox_col, group, s = 0.001, .n = 10, as.dist = FALSE, ...)
}
\arguments{
\item{...}{}

\item{.s}{Scale}
}
\description{
`euc-dist` gives a list of the clossesd regions of ore soruces to the sample
set based on euclidean distance.
`mass_frac_dist` returms a same with the mass fracsination correction applied.
}
