% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoprov_predict.R
\name{isoprov_predict}
\alias{isoprov_predict}
\title{Predict Isotope Provenance}
\usage{
isoprov_predict(x, model_list = NULL, .probablity = 0.95)
}
\arguments{
\item{x}{Matrix of liaendmembers object of pbisotope samples}

\item{model_list}{Model list gnerated by `train_data()`}

\item{.probablity}{Probablity between 0 and 1 of a Guess being correct. (Default = 0.95)}
}
\value{
data.frame object or list of data.frames
}
\description{
Predicuts Pb Isotope provencnace of a sample matrix in reference ot a xgboost trained list
}
\seealso{
train_data
}
