% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.liaendmembers}
\alias{summary.liaendmembers}
\title{Summary of LIA Endmember groups}
\usage{
\method{summary}{liaendmembers}(object, ...)
}
\arguments{
\item{object}{An object of class "liaendmembers"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing:

\item{Counts}{The number of observations in each group}
\item{Tolerance}{Tolerance values used in the calculation.}
\item{Clamp}{Clamping values used in the calculation.}
\item{Data}{A data frame of the lead isotope ratios that were used.}
}
\description{
Summary of Lead Isotope analysis 'liaendmemebr' object
}
\examples{
# Create object with class liaendmembers
data("tel_dor")
end_members <- endmembers(
        tel_dor,
        colnames(tel_dor),
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
# Print summary of the liaendmembers object
summary(end_members)
}
