% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.liaendmembers}
\alias{summary.liaendmembers}
\title{Summary of LIA Endmember groups}
\usage{
\method{summary}{liaendmembers}(x, ...)
}
\arguments{
\item{x}{An object of class "liaendmembers"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing:

\item{Counts}{The number of observations in each group}
\item{Tolerance}{The tolerance value used in the calculation.}
\item{Data}{A data frame of the lead isotope ratios that were used.}
}
\description{
Summary of Lead Isotope analysis 'liaendmemebr' object
}
\examples{
# Create object with class liaendmembers
data("dor")
end_members <- endmembers(
        dor,
        colnames(dor),
        tolerance = 0.01
)
# Prind summary of the liaendmembers object
summary.liaendmembers(end_members)
}
